+++ app/components/Navbar/ui/SearchBar.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Icons } from "@/app/ui/icons";
+import { Icons } from "@/app/utils/icons";
 
 export default function SearchBar() {
   return (
