+++ constants/categories.ts
@@ -1,3 +1,5 @@
+import { category } from "@/types";
+
 export const categories: category[] = [
   {
     name: "Bicycles",
