+++ package.json
@@ -16,8 +16,8 @@
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
-    "@uploadthing/react": "^6.2.0",
     "@tailwindcss/container-queries": "^0.1.1",
+    "@uploadthing/react": "^6.2.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "mongoose": "^8.1.0",
